--- src/main/java/kr/hhplus/be/server/domain/auth/constant/JwtTokenType.java
@@ -1,6 +1,0 @@
-package kr.hhplus.be.server.domain.auth.constant;
-
-public enum JwtTokenType {
-    ACCESS,
-    REFRESH
-}
