--- src/main/java/kr/hhplus/be/server/domain/auth/dto/request/ReIssueTokenRequest.java
@@ -1,9 +1,0 @@
-package kr.hhplus.be.server.domain.auth.dto.request;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record ReIssueTokenRequest(
-        @NotBlank String accessToken,
-        @NotBlank String refreshToken
-) {
-}
