--- src/main/java/kr/hhplus/be/server/domain/auth/dto/request/SignInRequest.java
@@ -1,9 +1,0 @@
-package kr.hhplus.be.server.domain.auth.dto.request;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record SignInRequest(
-        @NotBlank String name,
-        @NotBlank String password
-) {
-}
