--- src/main/java/kr/hhplus/be/server/domain/auth/dto/request/SignUpRequest.java
@@ -1,8 +1,0 @@
-package kr.hhplus.be.server.domain.auth.dto.request;
-
-public record SignUpRequest(
-        String email,
-        String name,
-        String password
-) {
-}
