--- src/main/java/kr/hhplus/be/server/domain/auth/dto/response/TokenPair.java
@@ -1,19 +1,0 @@
-package kr.hhplus.be.server.domain.auth.dto.response;
-
-import java.time.Instant;
-
-public record TokenPair(
-        Access access,
-        Refresh refresh
-) {
-
-    public record Access(
-            String token,
-            Instant expiredAt
-    ) {}
-
-    public record Refresh(
-            String token,
-            Instant expiredAt
-    ){}
-}
