--- src/main/java/kr/hhplus/be/server/domain/concert/constant/ConcertStatus.java
@@ -1,6 +1,0 @@
-package kr.hhplus.be.server.domain.concert.constant;
-
-public enum ConcertStatus {
-    ACTIVE,
-    INACTIVE,
-}
