--- src/main/java/kr/hhplus/be/server/domain/concert/exception/ConcertNotFoundException.java
@@ -1,9 +1,0 @@
-package kr.hhplus.be.server.domain.concert.exception;
-
-public class ConcertNotFoundException extends ConcertException {
-    private static final String MESSAGE = "콘서트를 찾을 수 없습니다.";
-
-    public ConcertNotFoundException() {
-        super(MESSAGE);
-    }
-}
