--- src/main/java/kr/hhplus/be/server/domain/user/constant/Role.java
@@ -1,6 +1,0 @@
-package kr.hhplus.be.server.domain.user.constant;
-
-public enum Role {
-    USER,
-    ADMIN
-}
