--- src/main/java/kr/hhplus/be/server/domain/user/exception/UserException.java
@@ -1,7 +1,0 @@
-package kr.hhplus.be.server.domain.user.exception;
-
-public abstract class UserException extends RuntimeException {
-    public UserException(String message) {
-        super(message);
-    }
-}
