+++ src/main/java/kr/hhplus/be/server/user/domain/auth/constant/JwtTokenType.java
@@ -1,0 +1,6 @@
+package kr.hhplus.be.server.user.domain.auth.constant;
+
+public enum JwtTokenType {
+    ACCESS,
+    REFRESH
+}
