+++ src/main/java/kr/hhplus/be/server/user/domain/auth/dto/request/ReIssueTokenRequest.java
@@ -1,0 +1,9 @@
+package kr.hhplus.be.server.user.domain.auth.dto.request;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record ReIssueTokenRequest(
+        @NotBlank String accessToken,
+        @NotBlank String refreshToken
+) {
+}
