+++ src/main/java/kr/hhplus/be/server/user/domain/auth/dto/request/SignInRequest.java
@@ -1,0 +1,9 @@
+package kr.hhplus.be.server.user.domain.auth.dto.request;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record SignInRequest(
+        String email,
+        @NotBlank String password
+) {
+}
