+++ src/main/java/kr/hhplus/be/server/user/domain/auth/dto/request/SignUpRequest.java
@@ -1,0 +1,8 @@
+package kr.hhplus.be.server.user.domain.auth.dto.request;
+
+public record SignUpRequest(
+        String email,
+        String name,
+        String password
+) {
+}
