+++ src/main/java/kr/hhplus/be/server/user/domain/auth/dto/response/TokenPair.java
@@ -1,0 +1,19 @@
+package kr.hhplus.be.server.user.domain.auth.dto.response;
+
+import java.time.Instant;
+
+public record TokenPair(
+        Access access,
+        Refresh refresh
+) {
+
+    public record Access(
+            String token,
+            Instant expiredAt
+    ) {}
+
+    public record Refresh(
+            String token,
+            Instant expiredAt
+    ) {}
+}
