+++ src/main/java/kr/hhplus/be/server/user/domain/auth/exception/AlreadySignOutTokenException.java
@@ -1,0 +1,9 @@
+package kr.hhplus.be.server.user.domain.auth.exception;
+
+public class AlreadySignOutTokenException extends AuthException {
+    public static final String MESSAGE = "이미 로그아웃 한 사용자입니다.";
+
+    public AlreadySignOutTokenException() {
+        super(MESSAGE);
+    }
+}
