+++ src/main/java/kr/hhplus/be/server/user/domain/concert/constant/ConcertStatus.java
@@ -1,0 +1,6 @@
+package kr.hhplus.be.server.user.domain.concert.constant;
+
+public enum ConcertStatus {
+    ACTIVE,
+    INACTIVE,
+}
