+++ src/main/java/kr/hhplus/be/server/user/domain/concert/exception/ConcertException.java
@@ -1,0 +1,9 @@
+package kr.hhplus.be.server.user.domain.concert.exception;
+
+import kr.hhplus.be.server._core.exception.ApiException;
+
+public abstract class ConcertException extends ApiException {
+    public ConcertException(String message) {
+        super(message);
+    }
+}
