+++ src/main/java/kr/hhplus/be/server/user/domain/concert/exception/ConcertNotFoundException.java
@@ -1,0 +1,9 @@
+package kr.hhplus.be.server.user.domain.concert.exception;
+
+public class ConcertNotFoundException extends ConcertException {
+    private static final String MESSAGE = "콘서트를 찾을 수 없습니다.";
+
+    public ConcertNotFoundException() {
+        super(MESSAGE);
+    }
+}
