+++ src/main/java/kr/hhplus/be/server/user/domain/user/constant/Role.java
@@ -1,0 +1,6 @@
+package kr.hhplus.be.server.user.domain.user.constant;
+
+public enum Role {
+    USER,
+    ADMIN
+}
