+++ src/main/java/kr/hhplus/be/server/user/domain/user/exception/UserException.java
@@ -1,0 +1,7 @@
+package kr.hhplus.be.server.user.domain.user.exception;
+
+public abstract class UserException extends RuntimeException {
+    public UserException(String message) {
+        super(message);
+    }
+}
